requirements.txt
google-cloud-bigquery==3.25.0
pandas==2.2.2
pyarrow==16.1.0
requests==2.32.3
python-dotenv==1.0.1
db-dtypes==1.4.3
sqlalchemy==2.0.20
sqlalchemy-bigquery==1.4.0